+++ src/pages/profile/editprofile.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import ProfileEditing from "../../components/Profile/ProfileEditing";
+import ProfileEditLayout from "../../components/Layout/ProfileEditLayout";
+
+const Editprofile = () => {
+  return (
+    <ProfileEditLayout>
+      <ProfileEditing />
+    </ProfileEditLayout>
+  );
+};
+
+export default Editprofile;
